tests: move TC_06_AppVMMixin from basic to misc
'basic' is ran on every storage pool and as such is a bit time
sensitive. Since tests in TC_06_AppVMMixin are getting longer (and not
really relevant for storage testing), move it to another file.
